tasks/requirements.txt
ujson==5.10.0
openpyxl==3.1.2
python-dotenv==1.0.1
aioclock==0.3.0
#easyocr==1.7.1
pytz==2024.1
termcolor==2.4.0
playwright==1.47.0
pydantic==2.5.3
tortoise-orm==0.21.5
asyncpg==0.29.0
cryptography==42.0.8
2captcha-python==1.5.0
httpx==0.27.0
pillow==10.4.0
anticaptchaofficial==1.0.61